Add consent read recipes
The DB now saves when you last updated your consent text, and when you
last read someone's consent text. If the update time is more recent than
the time you last read it, a red dot appears on the examine icon.

Some DB changes were made to support character-specific consent
settings. This isn't implemented yet, but it should save a migration for
if/when it does get implemented.